Castle implementation ++
Castle move now calls a separate move constructor, creates a move, then intentionally crashes the move constructor's try-catch to break out of it. Will assume castle is legal and simply post the move, and clear the old squares properly.

If called when it's illegal to castle will force a castle and kill it's own pieces in between. This is in AI because it considers the castle a Capture.